refactor(gazelle_cli): source_code_definition.dart
