+++ packages/gazelle_cli/lib/commands/codegen/generate_model_provider.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:code_builder/code_builder.dart' as cb;
 import 'package:dart_style/dart_style.dart';
 
-import 'class_definition.dart';
+import 'source_file_definition.dart';
 
 /// Represents the result of `GenerateModelProvider`.
 class GenerateModelProviderResult {
