+++ packages/gazelle_cli/lib/commands/codegen/source_file_definition.dart
